requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
spacy
python-docx
websockets
requests
